pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "crispy-bulma>=0.12.0",
    "django>=5.2.6",
    "django-crispy-forms>=2.4",
    "django-feather>=0.3.0",
    "django-htmx>=1.26.0",
    "django-simple-bulma>=2.7.1",
    "granian>=2.5.4",
    "pillow>=12.0.0",
    "pillow-avif-plugin>=1.5.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "granian[reload]>=2.5.4",
    "taskipy>=1.14.1",
    "isort>=7.0.0",
]

[tool.ruff]
exclude = [
    "venv/",
    ".venv/",
    "env/",
    ".env/",
    "build/",
    "dist/",
    "*.egg-info/",
    "*.db",
    "*.sqlite3",
    "*.log",
    "aegis_share/models.py",
    "aegis_share/migrations/",
    "test_tmp_*",
]

[tool.taskipy.tasks]
fmt = 'isort . && ruff check --fix && ruff format'
server = 'granian mysite.asgi:application  --interface asgi --reload --host 0.0.0.0 --port 8000'
migrate = 'uv run manage.py makemigrations && uv run manage.py migrate'
